requirements.txt
python-telegram-bot
apscheduler==3.10.4
httpx
pytz
six
